优化代码，增加了epoll驱动的写处理函数。

+++ net/xmn_socket.cc
@@ -760,7 +760,7 @@ void *XMNSocket::SendDataThread(void *pthreadinfo)
     XMNConnSockInfo *pconnsockinfo = nullptr;
     XMNMemory *pmemory = (XMNMemory *)SingletonBase<XMNMemory>::GetInstance();
     int err = 0;
-    int sendsize = 0;
+    ssize_t sendsize = 0;
 
     while (!g_isquit)
     {
